chore: split snapshotservice
